+++ src/main/java/model/piece/Horse.java
@@ -1,11 +1,9 @@
-package model.piece.normal;
+package model.piece;
 
 import java.util.List;
 
 import model.Position;
 import model.Team;
-import model.piece.Piece;
-import model.piece.PieceType;
 
 public class Horse extends Piece {
