+++ src/main/java/model/piece/movement/DefaultMovement.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 import model.Position;
 import model.Team;
-import model.piece.BoardSearcher;
+import model.board.BoardSearcher;
 import model.piece.Piece;
 
 public class DefaultMovement implements Movement {
