+++ src/main/java/model/piece/movement/PalaceMovement.java
@@ -8,7 +8,7 @@ import org.jetbrains.annotations.Nullable;
 
 import model.Position;
 import model.Team;
-import model.piece.BoardSearcher;
+import model.board.BoardSearcher;
 import model.piece.Piece;
 
 public class PalaceMovement implements Movement {
