+++ src/test/java/model/piece/PieceMoveTest.java
@@ -10,8 +10,6 @@ import org.junit.jupiter.api.Test;
 import model.Position;
 import model.Team;
 import model.board.Board;
-import model.piece.normal.Elephant;
-import model.piece.normal.Pawn;
 
 class PieceMoveTest {
